Separate plotter code from fitter code
Move common code to utility wrapper for MPB and have build plot
with different requirements for different services
(wip)
