+++ .gitignore
@@ -78,3 +78,7 @@ sketch
 
 # Pre commit configuration
 .pre-commit-config.yaml
+
+# Custom
+tmp/
+bin/
