+++ services/plotter/service.py
@@ -1,9 +1,9 @@
 from pandas import DataFrame
 
-from .plotter import build_plot
 from .._models.options import Options
 from .._utils.preprocessor import validate_dataframe, validate_latex
 from .._utils.log import logger
+from .._utils.mpb_wrap import set_func, set_plots, set_configurations
 
 
 def service(data: DataFrame, options: Options, func: str) -> bytes:
@@ -19,6 +19,16 @@ def service(data: DataFrame, options: Options, func: str) -> bytes:
     # TODO: convert latex to lambda code
 
     logger.debug("Building plot for given inputs.")
-    plot = build_plot(data, options, func)
+    
+    uid = uuid4()
+    fig: Figure = figure(uid)
+    axes: Axes = fig.add_axes([0.1, 0.1, 0.8, 0.8])
 
-    return plot
+    set_func(axes, func)
+    set_plots(axes, data, options.plots)
+    set_configurations(axes, options.figure)
+
+    buf = BytesIO()
+    fig.savefig(buf, format="png")
+
+    return buf.getvalue()
